Move createExistingCardMethod to custom hook
