+++ client/my-sites/checkout/checkout/composite-checkout/use-create-payment-methods.js
@@ -9,6 +9,7 @@ import {
 	createFullCreditsMethod,
 	createFreePaymentMethod,
 	createApplePayMethod,
+	createExistingCardMethod,
 	registerStore,
 	defaultRegistry,
 } from '@automattic/composite-checkout';
@@ -31,6 +32,7 @@ import {
 	WordPressFreePurchaseLabel,
 	WordPressFreePurchaseSummary,
 	submitApplePayPayment,
+	submitExistingCardPayment,
 } from '../composite-checkout-payment-methods';
 
 const debug = debugFactory( 'calypso:composite-checkout:use-create-payment-methods' );
@@ -265,4 +267,53 @@ export function useCreateApplePay( {
 		isApplePayAvailable,
 	] );
 	return applePayMethod;
+}
+
+export function useCreateExistingCards( {
+	onlyLoadPaymentMethods,
+	storedCards,
+	stripeConfiguration,
+} ) {
+	const shouldLoadExistingCardsMethods = onlyLoadPaymentMethods
+		? onlyLoadPaymentMethods.includes( 'existingCard' )
+		: true;
+	const existingCardMethods = useMemo( () => {
+		if ( ! shouldLoadExistingCardsMethods ) {
+			return [];
+		}
+		return storedCards.map( storedDetails =>
+			createExistingCardMethod( {
+				id: `existingCard-${ storedDetails.stored_details_id }`,
+				cardholderName: storedDetails.name,
+				cardExpiry: storedDetails.expiry,
+				brand: storedDetails.card_type,
+				last4: storedDetails.card,
+				stripeConfiguration,
+				submitTransaction: submitData => {
+					const pending = submitExistingCardPayment(
+						{
+							...submitData,
+							siteId: select( 'wpcom' )?.getSiteId?.(),
+							storedDetailsId: storedDetails.stored_details_id,
+							paymentMethodToken: storedDetails.mp_ref,
+							paymentPartnerProcessorId: storedDetails.payment_partner,
+							domainDetails: getDomainDetails( select ),
+						},
+						wpcomTransaction
+					);
+					// save result so we can get receipt_id and failed_purchases in getThankYouPageUrl
+					pending.then( result => {
+						debug( 'saving transaction response', result );
+						dispatch( 'wpcom' ).setTransactionResponse( result );
+					} );
+					return pending;
+				},
+				registerStore,
+				getCountry: () => select( 'wpcom' )?.getContactInfo?.()?.countryCode?.value,
+				getPostalCode: () => select( 'wpcom' )?.getContactInfo?.()?.postalCode?.value,
+				getSubdivisionCode: () => select( 'wpcom' )?.getContactInfo?.()?.state?.value,
+			} )
+		);
+	}, [ stripeConfiguration, storedCards, shouldLoadExistingCardsMethods ] );
+	return existingCardMethods;
 }
